requirements.txt
Flask==2.1.0
Flask-MySQLdb==1.0.1
PyYAML==6.0
mysqlclient==2.1.0